component-entrypoints
- `group-overflow`: добавлена точка входа `index.ts` (patch)
- `info-text`: добавлена точка входа `index.ts` (patch)
- `input`: добавлена точка входа `index.ts` (patch)
- `layout`: мелкие правки типов (patch)
- `link`: добавлена точка входа `index.ts` (patch)
- `modal-overlay`: добавлена точка входа `index.ts` (patch)
