+++ src/group-overflow/index.ts
@@ -1,0 +1,2 @@
+export type { GroupOverflowProps, RenderTail } from './types';
+export { GroupOverflow } from './group-overflow';
