+++ src/info-text/index.ts
@@ -1,0 +1,2 @@
+export type { InfoTextProps } from './types';
+export { InfoText } from './info-text';
