+++ src/input/index.ts
@@ -1,0 +1,2 @@
+export type { InputProps, InputType } from './types';
+export { Input } from './input';
