+++ src/layout/types.ts
@@ -1,7 +1,10 @@
-import { HTMLAttributes } from 'react';
-import { Breakpoint } from '../types';
+import type { HTMLAttributes, Ref } from 'react';
+import type { Breakpoint } from '../types';
 
 export interface LayoutProps extends HTMLAttributes<HTMLElement> {
   /** Список точек остановки на которых необходимо отключить ограничение ширины. */
   disabledOn?: Breakpoint[];
+
+  /** Реф корневого элемента. */
+  rootRef?: Ref<HTMLDivElement>;
 }
