+++ src/link/index.ts
@@ -1,0 +1,2 @@
+export type { LinkColor, LinkProps } from './types';
+export { Link } from './link';
