+++ src/modal-overlay/index.ts
@@ -1,0 +1,2 @@
+export type { ModalOverlayProps } from './types';
+export { ModalOverlay } from './modal-overlay';
